backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.2.1

# HTTP Client
httpx==0.27.0

# Database (Supabase Postgres)
sqlalchemy[asyncio]==2.0.30
psycopg[binary]==3.1.18

# Supabase - Using latest version with compatible dependencies
supabase==2.16.0

# Authentication & Security
python-jose[cryptography]==3.3.0

# Logging & Observability
structlog==24.1.0
google-cloud-logging==3.11.3

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
aiosqlite==0.19.0
